A C program to print the fibonacci series

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int x = 1, y = 1, n;
-	unsigned long long int z;
+	long int z;
 	
 	printf("%d, %d, ", x, y);
 
@@ -21,11 +21,11 @@ int main(void)
 		
 		if (n != 46)
 		{
-			printf("%lli, ", z);
+			printf("%li, ", z);
 		}
 		else
 		{
-			printf("%lli", z);
+			printf("%li", z);
 		}
 	}
 	printf("\n");
